requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.5
Flask_WTF==1.1.1
matplotlib==3.7.2
mlxtend==0.22.0
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
pandas==2.0.3
plotly==5.15.0
Pygments==2.15.1
pytest==7.4.0
python-dotenv==1.0.0
python_dateutil==2.8.2
scikit_learn==1.3.0
setuptools==66.0.0
setuptools==65.5.0
SQLAlchemy==2.0.19
statsmodels==0.14.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
treelib==1.6.4
Werkzeug==2.3.6
WTForms==3.0.1
